+++ src/main/java/org/jsapar/parse/cell/IntegerCellFactory.java
@@ -3,7 +3,6 @@ package org.jsapar.parse.cell;
 import org.jsapar.model.Cell;
 import org.jsapar.model.IntegerCell;
 import org.jsapar.text.Format;
-import org.jsapar.text.NumberFormat;
 
 import java.text.ParseException;
 import java.util.Locale;
@@ -22,6 +21,6 @@ public class IntegerCellFactory extends NumberCellFactory {
 
     @Override
     public Format makeFormat(Locale locale) {
-        return new NumberFormat( java.text.NumberFormat.getIntegerInstance(locale));
+        return Format.ofIntegerInstance(locale);
     }
 }