+++ src/main/java/org/jsapar/text/JavaTextFormat.java
@@ -3,10 +3,14 @@ package org.jsapar.text;
 import java.text.ParseException;
 import java.text.ParsePosition;
 
-public class JavaTextFormat<T> implements Format<T>{
+/**
+ * Formats and parses using a java.text.Format. Requires that all characters are parsed from the supplied value.
+ * @param <T>
+ */
+class JavaTextFormat<T> implements Format<T>{
     private java.text.Format format;
 
-    public JavaTextFormat(java.text.Format format) {
+    JavaTextFormat(java.text.Format format) {
         this.format = format;
     }